# perf: Fill buffer to minimize copies