```
goos: linux
goarch: amd64
pkg: github.com/ianlewis/runeio
cpu: Intel(R) Core(TM) i7-7567U CPU @ 3.50GHz
                     │   old.txt    │               new.txt                │
                     │    sec/op    │    sec/op     vs base                │
ReadRune-4             6.238n ± 10%   4.649n ±  9%  -25.47% (p=0.000 n=15)
ReadRuneUnbuffered-4   9.886n ±  7%   6.758n ±  7%  -31.64% (p=0.000 n=15)
ReadSmall-4            3.015µ ± 10%   1.988µ ± 12%  -34.06% (p=0.000 n=15)
ReadLarge-4            201.3µ ± 15%   123.8µ ± 11%  -38.51% (p=0.000 n=15)
NoCopySmall-4          250.4µ ±  1%   206.3µ ±  7%  -17.61% (p=0.000 n=15)
NoCopyLarge-4          7.780m ±  1%   6.648m ±  7%  -14.55% (p=0.000 n=15)
PeekSmall-4            3.527n ±  6%   3.275n ±  4%   -7.14% (p=0.000 n=15)
PeekLarge-4            3.526n ±  0%   3.276n ±  6%   -7.09% (p=0.000 n=15)
DiscardSmall-4         2.492µ ± 13%   1.759µ ±  7%  -29.41% (p=0.000 n=15)
DiscardLarge-4         181.8µ ±  8%   114.2µ ± 16%  -37.17% (p=0.000 n=15)
Fill-4                 36.42µ ±  6%   36.89µ ±  7%   +1.30% (p=0.016 n=15)
geomean                2.396µ         1.845µ        -23.01%
```

Fixes #52