@ianlewis 2023-11-11
I still don't quite get why some of these benchmarks are faster. They should be doing basically the same amount of work since the buffer was pre-filled.